requirements.txt
alabaster==0.7.12
aniso8601==3.0.2
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
commonmark==0.8.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.8
django-allauth==0.38.0
django-facebook==6.0.3
django-facebook-login==4.0.0
django-heroku==0.3.1
django-jsonfield==1.0.1
djangorestframework==3.9.1
docutils==0.14
future==0.17.1
graphene==2.1.3
graphene-django==2.1.0
graphql-core==2.1
graphql-relay==0.4.5
gunicorn==19.9.0
idna==2.8
imagesize==1.1.0
iso8601==0.1.12
Jinja2==2.10
jsonfield==2.0.2
MarkupSafe==1.1.0
oauthlib==3.0.1
packaging==19.0
promise==2.2.1
psycopg2==2.7.7
purepng==0.2.0
Pygments==2.3.1
PyJWT==1.7.1
pyparsing==2.3.1
python-dateutil==2.7.5
python-social-auth==0.2.1
python3-openid==3.1.0
pytz==2018.7
recommonmark==0.5.0
requests==2.21.0
requests-oauthlib==1.2.0
rinoh-typeface-dejavuserif==0.1.1
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.3.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
social-auth-app-django==3.1.0
social-auth-core==3.0.0
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
transitions==0.6.9
Unidecode==1.0.23
urllib3==1.24.1
whitenoise==4.1.2
